refactor: share client provider runtime without merging DOM applies
Default and extended client providers duplicated store init, history reapply, and storage/system effects. Extract that runtime into one hook and a shared storage helper, keep client-dom and extended-client-dom apply/bootstrap split, and raise next-provider budgets for the honest abstraction cost.
