+++ packages/themes/src/providers/extended-client-provider.tsx
@@ -1,36 +1,14 @@
 "use client";
 
-import {
-	type ReactElement,
-	useCallback,
-	useEffect,
-	useMemo,
-	useRef,
-	useSyncExternalStore,
-} from "react";
-import {
-	holdIfEqual,
-	sameStringList,
-	sameStringRecord,
-	sameThemeColor,
-} from "../core/config-equal.js";
+import { type ReactElement, useEffect, useMemo, useRef } from "react";
 import { ThemeContext } from "../core/context.js";
-import {
-	type AppliedThemeState,
-	applyExtendedThemeToDom,
-	getDomWindow,
-	readStoredTheme,
-	writeStoredTheme,
-} from "../core/extended-client-dom.js";
-import { subscribeHistoryReapply } from "../core/history-reapply.js";
+import { type AppliedThemeState, applyExtendedThemeToDom } from "../core/extended-client-dom.js";
 import type { ExtendedThemeProviderProps, SystemThemeMap } from "../core/extended-types.js";
-import { createThemeStore } from "../core/store.js";
 import { publishThemeChannel, subscribeThemeChannel } from "../core/sync.js";
-import { isThemeSelection, resolveDefaultTheme } from "../core/theme-validation.js";
+import { isThemeSelection } from "../core/theme-validation.js";
 import type { Attribute, DefaultTheme, ThemeContextValue } from "../core/types.js";
 import { useEffectEvent } from "../core/use-effect-event.js";
-
-const DEFAULT_THEMES: string[] = ["light", "dark"];
+import { type ClientThemeStableConfig, useClientThemeRuntime } from "./client-theme-runtime.js";
 
 type LastAppliedTheme = {
 	resolved: string;
@@ -71,7 +49,7 @@ function resolveSelection(
 
 export function ExtendedClientThemeProvider<Themes extends string = DefaultTheme>({
 	children,
-	themes = DEFAULT_THEMES as Themes[],
+	themes,
 	forcedTheme,
 	enableSystem = true,
 	defaultTheme,
@@ -79,11 +57,11 @@ export function ExtendedClientThemeProvider<Themes extends string = DefaultTheme
 	value: valueMap,
 	target = "html",
 	disableTransitionOnChange = false,
-	storage = "localStorage",
+	storage,
 	storageKey = "theme",
 	enableColorScheme = true,
 	themeColor,
-	followSystem = false,
+	followSystem,
 	onThemeChange,
 	initialTheme,
 	cookieOptions,
@@ -92,291 +70,90 @@ export function ExtendedClientThemeProvider<Themes extends string = DefaultTheme
 	systemThemeMap,
 	themeRoot,
 }: ExtendedThemeProviderProps<Themes>): ReactElement {
-	const resolvedDefault = resolveDefaultTheme(themes, enableSystem, defaultTheme);
-
-	const storeRef = useRef<ReturnType<typeof createThemeStore> | null>(null);
-	if (storeRef.current === null) {
-		storeRef.current = createThemeStore();
-	}
-	const store = storeRef.current;
 	const appliedThemeRef = useRef<AppliedThemeState | undefined>(undefined);
-	const lastAppliedRef = useRef<LastAppliedTheme | null>(null);
-	const themesRef = useRef(themes);
-	themesRef.current = holdIfEqual(themesRef.current, themes, sameStringList);
-	const stableThemes = themesRef.current;
-	const valueMapRef = useRef(valueMap);
-	valueMapRef.current = holdIfEqual(valueMapRef.current, valueMap, sameStringRecord);
-	const stableValueMap = valueMapRef.current;
-	const themeColorRef = useRef(themeColor);
-	themeColorRef.current = holdIfEqual(themeColorRef.current, themeColor, sameThemeColor);
-	const stableThemeColor = themeColorRef.current;
-	const {
-		getSnapshot,
-		setState: setStoreState,
-		setTheme: setStoreTheme,
-		setSystemTheme: setStoreSystemTheme,
-	} = store;
-
-	const { theme, systemTheme } = useSyncExternalStore(
-		store.subscribe,
-		store.getSnapshot,
-		store.getServerSnapshot,
-	);
-
-	const validForcedTheme =
-		forcedTheme && stableThemes.includes(forcedTheme) ? forcedTheme : undefined;
-	const selectedTheme = validForcedTheme ?? theme;
-	const resolvedTheme = selectedTheme
-		? (resolveSelection(
-				selectedTheme,
-				systemTheme,
-				systemThemeMap as SystemThemeMap<string> | undefined,
-			) as Themes | undefined)
-		: undefined;
 	const channel = `${storage ?? "localStorage"}:${storageKey}:${target}`;
 
-	const onThemeChangeEvent = useEffectEvent((next: Themes) => {
-		onThemeChange?.(next);
-	});
-
-	const applyToDomEvent = useEffectEvent((resolved: string) => {
-		const last: LastAppliedTheme = {
+	const {
+		theme,
+		systemTheme,
+		resolvedTheme,
+		validForcedTheme,
+		stableThemes,
+		setTheme,
+		getSnapshot,
+		setStoreTheme,
+		applyResolved,
+	} = useClientThemeRuntime<Themes, LastAppliedTheme>({
+		themes,
+		forcedTheme,
+		enableSystem,
+		defaultTheme,
+		value: valueMap,
+		storage,
+		storageKey,
+		followSystem,
+		onThemeChange,
+		initialTheme,
+		cookieOptions,
+		onStorageError,
+		themeColor,
+		keepSystemSelection: !!systemThemeMap && !isDirectSystemMap(systemThemeMap),
+		resolveTheme: (selection, systemThemeValue) =>
+			resolveSelection(
+				selection,
+				systemThemeValue,
+				systemThemeMap as SystemThemeMap<string> | undefined,
+			),
+		createLast: (resolved, config: ClientThemeStableConfig) => ({
 			resolved,
 			attribute,
-			themes: stableThemes,
-			valueMap: stableValueMap,
+			themes: config.themes,
+			valueMap: config.valueMap,
 			target,
 			disableTransitionOnChange,
 			enableColorScheme,
-			themeColor: stableThemeColor,
+			themeColor: config.themeColor,
 			themeRoot,
-		};
-		appliedThemeRef.current = applyExtendedThemeToDom({
-			...last,
-			previous: appliedThemeRef.current,
-		});
-		lastAppliedRef.current = last;
-	});
-
-	const initializeEvent = useEffectEvent(() => {
-		const domWindow = getDomWindow();
-		if (!domWindow) return;
-		const mq =
-			enableSystem && typeof domWindow.matchMedia === "function"
-				? domWindow.matchMedia("(prefers-color-scheme: dark)")
-				: null;
-		const system = mq ? (mq.matches ? "dark" : "light") : undefined;
-		let initial: Themes | "system";
-
-		// Forced theme short-circuits init and must not persist to storage.
-		if (validForcedTheme) {
-			initial = validForcedTheme;
-		} else if (initialTheme && isThemeSelection(initialTheme, themes, enableSystem)) {
-			initial = initialTheme;
-			writeStoredTheme(
-				storage,
-				storageKey,
-				String(initialTheme),
-				cookieOptions,
-				onStorageError,
-			);
-		} else {
-			const stored = readStoredTheme(storage, storageKey, onStorageError);
-			initial =
-				!followSystem && stored && isThemeSelection(stored, themes, enableSystem)
-					? (stored as Themes | "system")
-					: resolvedDefault;
-		}
-
-		setStoreState({ theme: initial, systemTheme: system });
-	});
-
-	const handleSystemChangeEvent = useEffectEvent((next: "light" | "dark") => {
-		setStoreSystemTheme(next);
-		const current = getSnapshot().theme;
-		if (current === "system" || current === undefined || followSystem) {
-			const followsVariant =
-				followSystem && !!systemThemeMap && !isDirectSystemMap(systemThemeMap);
-			if (followSystem && !followsVariant) setStoreTheme("system");
-			const resolved =
-				resolveSelection(
-					followsVariant
-						? (current ?? resolvedDefault)
-						: followSystem
-							? "system"
-							: (current ?? "system"),
-					next,
-					systemThemeMap as SystemThemeMap<string> | undefined,
-				) ?? next;
-			applyToDomEvent(resolved);
-			onThemeChangeEvent(next as Themes);
-		}
-	});
-
-	// oxlint-disable-next-line react-hooks/exhaustive-deps -- applyToDomEvent is an effect event.
-	const setTheme = useCallback(
-		(
-			next:
-				| Themes
-				| "system"
-				| ((current: Themes | "system" | undefined) => Themes | "system"),
-		) => {
-			if (validForcedTheme) return;
-
-			const current = getSnapshot().theme as Themes | "system" | undefined;
-			const newTheme = typeof next === "function" ? next(current) : next;
-			if (!isThemeSelection(newTheme, stableThemes, enableSystem)) return;
-			const resolved = resolveSelection(
-				newTheme,
-				getSnapshot().systemTheme,
-				systemThemeMap as SystemThemeMap<string> | undefined,
-			);
-
-			setStoreTheme(newTheme);
-			// oxlint-disable-next-line react-hooks/rules-of-hooks -- shared apply path; setTheme stays a public callback.
-			if (resolved) applyToDomEvent(resolved);
-			onThemeChange?.(newTheme as Themes);
-			writeStoredTheme(storage, storageKey, newTheme, cookieOptions, onStorageError);
-			if (enableSameDocumentSync && storage !== "none")
-				publishThemeChannel(channel, newTheme);
-		},
-		[
-			validForcedTheme,
-			stableThemes,
-			enableSystem,
-			systemThemeMap,
-			storage,
-			storageKey,
-			cookieOptions,
-			onStorageError,
-			channel,
-			enableSameDocumentSync,
-			getSnapshot,
-			setStoreTheme,
-			onThemeChange,
-		],
-	);
-
-	// oxlint-disable-next-line react-hooks/exhaustive-deps -- effect events are intentionally non-reactive.
-	useEffect(() => {
-		initializeEvent();
-		const domWindow = getDomWindow();
-		if (!domWindow) return;
-		return subscribeHistoryReapply(domWindow, () => {
-			const last = lastAppliedRef.current;
-			if (last) applyToDomEvent(last.resolved);
-		});
-	}, []);
-
-	// setTheme / system / storage already write the DOM. Re-apply only when
-	// resolvedTheme or DOM config changed from outside those paths.
-	// oxlint-disable-next-line react-hooks/exhaustive-deps -- effect events are intentionally non-reactive.
-	useEffect(() => {
-		if (!resolvedTheme) return;
-		const last = lastAppliedRef.current;
-		if (
-			last &&
-			last.resolved === resolvedTheme &&
+		}),
+		sameLast: (last, resolved, config) =>
+			last.resolved === resolved &&
 			last.attribute === attribute &&
-			last.themes === stableThemes &&
-			last.valueMap === stableValueMap &&
+			last.themes === config.themes &&
+			last.valueMap === config.valueMap &&
 			last.target === target &&
 			last.disableTransitionOnChange === disableTransitionOnChange &&
 			last.enableColorScheme === enableColorScheme &&
-			last.themeColor === stableThemeColor &&
-			last.themeRoot === themeRoot
-		) {
-			return;
-		}
-		applyToDomEvent(resolvedTheme);
-	}, [
-		resolvedTheme,
-		attribute,
-		stableThemes,
-		stableValueMap,
-		target,
-		disableTransitionOnChange,
-		enableColorScheme,
-		stableThemeColor,
-		themeRoot,
-	]);
+			last.themeColor === config.themeColor &&
+			last.themeRoot === themeRoot,
+		applyLast: (last) => {
+			appliedThemeRef.current = applyExtendedThemeToDom({
+				...last,
+				previous: appliedThemeRef.current,
+			});
+		},
+		afterSetTheme: (newTheme) => {
+			if (enableSameDocumentSync && storage !== "none") {
+				publishThemeChannel(channel, newTheme);
+			}
+		},
+	});
 
-	// oxlint-disable-next-line react-hooks/exhaustive-deps -- effect events are intentionally non-reactive.
-	useEffect(() => {
-		const domWindow = getDomWindow();
-		if (!enableSystem || !domWindow || typeof domWindow.matchMedia !== "function") return;
-		const mq = domWindow.matchMedia("(prefers-color-scheme: dark)");
-		setStoreSystemTheme(mq.matches ? "dark" : "light");
-		const handler = (event: MediaQueryListEvent) => {
-			handleSystemChangeEvent(event.matches ? "dark" : "light");
-		};
-		mq.addEventListener?.("change", handler);
-		return () => mq.removeEventListener?.("change", handler);
-	}, [enableSystem, setStoreSystemTheme]);
-
-	// oxlint-disable-next-line react-hooks/exhaustive-deps -- effect events are intentionally non-reactive.
-	useEffect(() => {
-		const domWindow = getDomWindow();
-		if (!domWindow) return;
-		if (
-			followSystem ||
-			storage === "none" ||
-			storage === "sessionStorage" ||
-			storage === "cookie"
-		)
-			return;
-
-		const handler = (event: StorageEvent) => {
-			if (event.storageArea !== localStorage || event.key !== storageKey) return;
-			const newTheme = event.newValue ?? resolvedDefault;
-			if (!isThemeSelection(newTheme, themes, enableSystem)) return;
-			const resolved = resolveSelection(
-				newTheme,
-				getSnapshot().systemTheme,
-				systemThemeMap as SystemThemeMap<string> | undefined,
-			);
-			setStoreTheme(newTheme);
-			if (!validForcedTheme && resolved) applyToDomEvent(resolved);
-		};
-		domWindow.addEventListener("storage", handler);
-		return () => domWindow.removeEventListener("storage", handler);
-	}, [
-		followSystem,
-		storage,
-		storageKey,
-		resolvedDefault,
-		themes,
-		enableSystem,
-		systemThemeMap,
-		validForcedTheme,
-		getSnapshot,
-		setStoreTheme,
-	]);
+	const applyChannelThemeEvent = useEffectEvent((newTheme: string) => {
+		if (!isThemeSelection(newTheme, themes, enableSystem)) return;
+		setStoreTheme(newTheme);
+		const resolved = resolveSelection(
+			newTheme,
+			getSnapshot().systemTheme,
+			systemThemeMap as SystemThemeMap<string> | undefined,
+		);
+		if (!validForcedTheme && resolved) applyResolved(resolved);
+	});
 
 	// oxlint-disable-next-line react-hooks/exhaustive-deps -- effect events are intentionally non-reactive.
 	useEffect(() => {
 		if (!enableSameDocumentSync || storage === "none") return;
-		return subscribeThemeChannel(channel, (newTheme) => {
-			if (!isThemeSelection(newTheme, themes, enableSystem)) return;
-			setStoreTheme(newTheme);
-			const resolved = resolveSelection(
-				newTheme,
-				getSnapshot().systemTheme,
-				systemThemeMap as SystemThemeMap<string> | undefined,
-			);
-			if (!validForcedTheme && resolved) applyToDomEvent(resolved);
-		});
-	}, [
-		enableSameDocumentSync,
-		storage,
-		channel,
-		themes,
-		enableSystem,
-		systemThemeMap,
-		validForcedTheme,
-		getSnapshot,
-		setStoreTheme,
-	]);
+		return subscribeThemeChannel(channel, applyChannelThemeEvent);
+	}, [enableSameDocumentSync, storage, channel]);
 
 	const contextValue = useMemo(
 		(): ThemeContextValue<string> => ({
